requirements.txt
accelerate==1.12.0
aiofiles==25.1.0
annotated-doc==0.0.4
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.12.1
asttokens==3.0.1
attrs==25.4.0
backoff==2.2.1
beautifulsoup4==4.14.3
brotli==1.2.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
colorlog==6.10.1
comm==0.2.3
contourpy==1.3.3
cryptography==46.0.5
cssselect2==0.8.0
cycler==0.12.1
dataclasses-json==0.6.7
debugpy==1.8.20
decorator==5.2.1
Deprecated==1.3.1
dill==0.4.1
distro==1.9.0
docling==2.73.0
docling-core==2.64.0
docling-ibm-models==3.11.0
docling-parse==4.7.3
edgartools==5.15.2
effdet==0.4.1
emoji==2.15.0
et_xmlfile==2.0.0
executing==2.2.1
Faker==40.4.0
filelock==3.20.3
filetype==1.2.0
flatbuffers==25.12.19
fonttools==4.61.1
fsspec==2026.2.0
google-api-core==2.29.0
google-auth==2.48.0
google-cloud-vision==3.12.1
googleapis-common-protos==1.72.0
greenlet==3.3.1
grpcio==1.78.0
grpcio-status==1.78.0
h11==0.16.0
h2==4.3.0
hf-xet==1.2.0
hpack==4.1.0
html5lib==1.1
httpcore==1.0.9
httpx==0.28.1
httpxthrottlecache==0.3.0
huggingface_hub==0.36.2
humanize==4.15.0
hyperframe==6.1.0
idna==3.11
ipykernel==7.2.0
ipython==9.10.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
jiter==0.13.0
joblib==1.5.3
jsonlines==4.0.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonref==1.1.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
jupyter_client==8.8.0
jupyter_core==5.9.1
kiwisolver==1.4.9
langchain-core==1.2.12
langdetect==1.0.9
langgraph==1.0.8
langgraph-checkpoint==4.0.0
langgraph-prebuilt==1.0.7
langgraph-sdk==0.3.5
langsmith==0.7.1
latex2mathml==3.78.1
llvmlite==0.46.0
lxml==6.0.2
markdown-it-py==4.0.0
marko==2.2.2
MarkupSafe==3.0.3
marshmallow==3.26.2
matplotlib==3.10.8
matplotlib-inline==0.2.1
mdurl==0.1.2
ml_dtypes==0.5.4
mpire==2.10.2
mpmath==1.3.0
multiprocess==0.70.19
mypy_extensions==1.1.0
nest-asyncio==1.6.0
networkx==3.6.1
nltk==3.9.2
numba==0.63.1
numpy==2.3.5
olefile==0.47
omegaconf==2.3.0
onnx==1.20.1
onnxruntime==1.24.1
openai==2.20.0
opencv-python==4.13.0.92
openpyxl==3.1.5
orjson==3.11.7
ormsgpack==1.12.2
packaging==26.0
pandas==2.3.3
parso==0.8.6
pdf2image==1.17.0
pdfminer.six==20260107
pi_heif==1.2.0
pikepdf==10.3.0
pillow==11.3.0
platformdirs==4.7.0
playwright==1.58.0
pluggy==1.6.0
polyfactory==3.2.0
portalocker==3.2.0
prompt_toolkit==3.0.52
proto-plus==1.27.1
protobuf==6.33.5
psutil==7.2.2
pure_eval==0.2.3
pyarrow==23.0.0
pyasn1==0.6.2
pyasn1_modules==0.4.2
pyclipper==1.4.0
pycocotools==2.0.11
pycparser==3.0
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
pydyf==0.12.1
pyee==13.0.0
Pygments==2.19.2
pylatexenc==2.10
pyparsing==3.3.2
pypdf==6.7.0
pypdfium2==5.4.0
pyphen==0.17.2
pyrate-limiter==3.9.0
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.2.1
python-iso639==2026.1.31
python-magic==0.4.27
python-multipart==0.0.22
python-oxmsg==0.0.2
python-pptx==1.0.2
pytz==2025.2
pywin32==311
PyYAML==6.0.3
pyzmq==27.1.0
qdrant-client==1.16.2
rank-bm25==0.2.2
RapidFuzz==3.14.3
rapidocr==3.6.0
referencing==0.37.0
regex==2026.1.15
requests==2.32.5
requests-toolbelt==1.0.0
rich==14.3.2
rpds-py==0.30.0
rsa==4.9.1
rtree==1.4.1
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.17.0
sec-api==1.0.32
sec-edgar-downloader==5.1.0
semchunk==2.2.2
sentence-transformers==5.2.2
setuptools==82.0.0
shapely==2.1.2
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.8.3
stack-data==0.6.3
stamina==25.2.0
sympy==1.14.0
tabulate==0.9.0
tenacity==9.1.4
textdistance==4.6.3
threadpoolctl==3.6.0
timm==1.0.24
tinycss2==1.5.1
tinyhtml5==2.0.0
tokenizers==0.22.2
torch==2.10.0+cu126
torchvision==0.25.0
tornado==6.5.4
tqdm==4.67.3
traitlets==5.14.3
transformers==4.57.6
tree-sitter==0.25.2
tree-sitter-c==0.24.1
tree-sitter-javascript==0.25.0
tree-sitter-python==0.25.0
tree-sitter-typescript==0.23.2
typer==0.21.2
typer-slim==0.23.0
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
Unidecode==1.4.0
unstructured-client==0.42.10
unstructured.pytesseract==0.3.15
unstructured_inference==1.2.0
urllib3==2.6.3
uuid_utils==0.14.0
wcwidth==0.6.0
weasyprint==68.1
webencodings==0.5.1
wrapt==2.1.1
xlsxwriter==3.2.9
xxhash==3.6.0
zopfli==0.4.0
zstandard==0.25.0
